refactoring charge values, formatting fixes

+++ Src/ToolType.java
@@ -1,17 +1,13 @@
 package Src;
-
-
-import java.math.BigDecimal;
-
 public class ToolType {
 
     private String type;
-    private BigDecimal dailyCharge;
+    private double dailyCharge;
     private boolean weekdayCharge;
     private boolean weekendCharge;
     private boolean holidayCharge;
 
-    public ToolType(String type, BigDecimal dailyCharge, boolean weekdayCharge, boolean weekendCharge, boolean holidayCharge) {
+    public ToolType(String type, double dailyCharge, boolean weekdayCharge, boolean weekendCharge, boolean holidayCharge) {
         this.type = type;
         this.dailyCharge = dailyCharge;
         this.weekdayCharge = weekdayCharge;
@@ -27,11 +23,11 @@ public class ToolType {
         this.type = type;
     }
 
-    public BigDecimal getDailyCharge() {
+    public double getDailyCharge() {
         return dailyCharge;
     }
 
-    public void setDailyCharge(BigDecimal dailyCharge) {
+    public void setDailyCharge(double dailyCharge) {
         this.dailyCharge = dailyCharge;
     }
